+++ Lib/test/test_decimal/__main__.py
@@ -9,11 +9,13 @@ def execute_tests(arith, verbose, todo_tests=None, debug=None):
     is enabled in regrtest.py
     """
 
-    module = getattr(__import__("test.test_decimal"), "test_decimal")
+    module = __import__("test.test_decimal")
+    module = module.test_decimal
     module.TEST_ALL = arith
     module.TODO_TESTS = todo_tests
     module.DEBUG = debug
     unittest.main(module, verbosity=2 if verbose else 1, exit=False, argv=[__name__])
+
 
 p = optparse.OptionParser("test_decimal [--debug] [{--skip | test1 [test2 [...]]}]")
 p.add_option('--debug', '-d', action='store_true', help='shows the test number and context before each test')
